feat: Add midsection of loginview
