+++ FlowUs/FlowUsApp.swift
@@ -32,7 +32,10 @@ struct FlowUsApp: App {
                 case .Welcome:
                     return AnyView(WelcomeView().navigationBarHidden(true))
                 case .Login:
-                    return AnyView(LoginView().navigationBarHidden(true).navigationBarBackButtonHidden(true))
+                    return AnyView(LoginView()
+                        .transition(.asymmetric(insertion: .slide, removal: .opacity))
+                        .navigationBarHidden(true)
+                        .navigationBarBackButtonHidden(true))
                 case .Register:
                     return AnyView(RegisterView().navigationBarHidden(true).navigationBarBackButtonHidden(true))
                 }
