+++ FlowUs/Views/shared/action/BackBar.swift
@@ -18,7 +18,7 @@ struct BackBar: View {
             shadowColor: .pink, width: 32, height: 32)
                 .padding(.leading, 8)
             Spacer()
-        }.padding(.top, -8)
+        }.padding(.top, UIDevice.current.hasNotch ? -8 : 0)
     }
 }
